Extract quic tests
