+++ rustls/tests/runners/api.rs
@@ -21,6 +21,8 @@ mod tests_with_ring {
     mod ffdhe;
     #[path = "../io.rs"]
     mod io;
+    #[path = "../quic.rs"]
+    mod quic;
     #[path = "../rpk.rs"]
     mod rpk;
     #[path = "../server_cert_verifier.rs"]
@@ -46,6 +48,8 @@ mod tests_with_aws_lc_rs {
     mod ffdhe;
     #[path = "../io.rs"]
     mod io;
+    #[path = "../quic.rs"]
+    mod quic;
     #[path = "../rpk.rs"]
     mod rpk;
     #[path = "../server_cert_verifier.rs"]
